Restructure of source files
