+++ src/__init__.py
@@ -1,0 +1,1 @@
+# src package for board game concept source files
